Refactoring Crawler hierarchy Close #37
